refactor: move register and storage clients to new namespace
