+++ src/Altinn.App.Core/Extensions/ServiceCollectionExtensions.cs
@@ -1,8 +1,8 @@
 using System;
 
 using Altinn.App.Core.Implementation;
-using Altinn.App.Core.Infrastructure.Register.HttpClients;
-using Altinn.App.Core.Infrastructure.Storage.HttpClients;
+using Altinn.App.Core.Infrastructure.Clients.Register;
+using Altinn.App.Core.Infrastructure.Clients.Storage;
 using Altinn.App.Core.Interface;
 using Altinn.App.PlatformServices.Implementation;
 using Altinn.App.PlatformServices.Interface;
