+++ src/Altinn.App.Core/Infrastructure/Clients/Register/RegisterERClient.cs
@@ -10,7 +10,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
-namespace Altinn.App.Core.Infrastructure.Register.HttpClients
+namespace Altinn.App.Core.Infrastructure.Clients.Register
 {
     /// <summary>
     /// A client for retrieving ER data from Altinn Platform.
